Encapsulate user trace rendering
