+++ ydb/core/kqp/session_actor/kqp_user_facing_tracing.h
@@ -5,6 +5,8 @@
 #include <ydb/core/protos/kqp.pb.h>
 #include <ydb/library/actors/wilson/wilson_trace.h>
 
+#include "kqp_query_stats.h"
+
 #include <util/generic/string.h>
 
 #include <optional>
@@ -26,6 +28,13 @@ constexpr size_t MaxUserFacingSpansPerQuery = 1000;
 TInstant MapUserFacingSessionStart(TInstant localStart, TInstant originSentAt,
     const google::protobuf::RepeatedPtrField<NKikimrKqp::TProxyRequestHop>& proxyHops);
 
+struct TUserFacingQueryDescription {
+    TString DisplayName;
+    TString Operation;
+};
+
+struct TUserFacingQuerySnapshot;
+
 class TUserFacingTraceContext {
 public:
     TUserFacingTraceContext(NWilson::TTraceId traceId, TInstant sessionStart,
@@ -37,6 +46,26 @@ public:
     TUserFacingTraceContext(TUserFacingTraceContext&&) = default;
     TUserFacingTraceContext& operator=(TUserFacingTraceContext&&) = default;
 
+    const TExecutionDiagnosticsPolicy& GetDiagnosticsPolicy() const;
+    void StartAdmission(TInstant at = TInstant::Now());
+    void FinishAdmission(Ydb::StatusIds::StatusCode status, TInstant at = TInstant::Now());
+    void MarkExecutionDelegated();
+
+    void BeginCompile(TInstant at = TInstant::Now());
+    void RecordCompileCacheHit(Ydb::StatusIds::StatusCode status,
+        TInstant at = TInstant::Now());
+    void FinishCompile(bool fromCache, Ydb::StatusIds::StatusCode status,
+        std::shared_ptr<const TCompileDiagnostics> dependencies,
+        std::optional<TCompileActorDiagnostic> actor, TInstant at = TInstant::Now());
+    void FinishSplit(Ydb::StatusIds::StatusCode status, TInstant at = TInstant::Now());
+
+    void AddExecutions(std::vector<TExecutionTraceSnapshot>& source,
+        const TExecutionTraceTotals& totals, size_t sourceDropped);
+    void UpdateQueryDescription(const TUserFacingQueryDescription& description);
+    TUserFacingQuerySnapshot DetachSnapshot(TKqpQueryState& state, bool success,
+        const TString& statusCode, NKikimrKqp::TEvQueryResponse* response);
+
+private:
     NWilson::TTraceId TraceId;
     TExecutionDiagnosticsPolicy DiagnosticsPolicy;
     TInstant SessionStart;
@@ -56,6 +85,42 @@ public:
     TExecutionTraceTotals ExecutionTraceTotals;
     size_t ExecutionTracesDropped = 0;
 };
+
+// Detached from the session actor and consumed asynchronously by the renderer.
+struct TUserFacingQuerySnapshot {
+    NWilson::TTraceId TraceId;
+    TString RootName;
+    TString Operation;
+    TString QueryText;
+    TInstant RootEnd;
+    TInstant StartTime;
+    std::vector<NKikimrKqp::TProxyRequestHop> ProxyRequestHops;
+    TInstant AdmissionStartedAt;
+    TInstant AdmissionFinishedAt;
+    Ydb::StatusIds::StatusCode AdmissionStatus = Ydb::StatusIds::STATUS_CODE_UNSPECIFIED;
+    TString PoolId;
+    TKqpQueryStats QueryStats;
+    std::vector<TExecutionTraceSnapshot> ExecutionTraces;
+    TExecutionTraceTotals ExecutionTraceTotals;
+    size_t ExecutionTracesDropped = 0;
+    std::vector<TCompileAttemptDiagnostic> CompileAttempts;
+    size_t CompileAttemptsDropped = 0;
+    bool ExecutionDelegated = false;
+    bool Success = false;
+    TString StatusCode;
+};
+
+struct TRejectedUserFacingQuerySnapshot {
+    NWilson::TTraceId TraceId;
+    TString QueryText;
+    std::vector<NKikimrKqp::TProxyRequestHop> ProxyRequestHops;
+    TInstant RejectedAt;
+    Ydb::StatusIds::StatusCode Status = Ydb::StatusIds::STATUS_CODE_UNSPECIFIED;
+};
+
+NActors::IActor* CreateUserFacingTraceRendererActor(TUserFacingQuerySnapshot snapshot);
+NActors::IActor* CreateRejectedUserFacingTraceRendererActor(
+    TRejectedUserFacingQuerySnapshot snapshot);
 
 TTimeWindow FitUserFacingRemoteWindow(TTimeWindow window, const TTimeWindow& parent);
 
@@ -95,11 +160,6 @@ private:
     size_t Remaining_;
     ui8 Verbosity_;
     ui64 Dropped_ = 0;
-};
-
-struct TUserFacingQueryDescription {
-    TString DisplayName;
-    TString Operation;
 };
 
 TUserFacingQueryDescription DescribeUserFacingQuery(const TKqpQueryState& state);
