+++ ydb/core/kqp/session_actor/ya.make
@@ -7,6 +7,7 @@ SRCS(
     kqp_response.cpp
     kqp_session_actor.cpp
     kqp_user_facing_query_description.cpp
+    kqp_user_facing_trace_renderer.cpp
     kqp_user_facing_tracing.cpp
     kqp_temp_tables_manager.cpp
     kqp_worker_actor.cpp
